# docs/Api-Documentation.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "E-Commerce Ordering & Payment API",
        "description": "Backend for managing users, products, orders, and payments with support for multiple payment providers (Stripe, bKash).",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/users/me/orders": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Read Own Orders",
                "operationId": "read_own_orders_api_v1_users_me_orders_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/OrderRead"
                                    },
                                    "type": "array",
                                    "title": "Response Read Own Orders Api V1 Users Me Orders Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
        "/api/v1/orders": {
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Create Order",
                "description": "Place an order for one or more products. Totals are computed server-side;\nthe order starts `pending` and reserves no stock.",
                "operationId": "create_order_api_v1_orders_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/v1/orders/{order_id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get Order",
                "operationId": "get_order_api_v1_orders__order_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Order Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{order_id}/cancel": {
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Cancel Order",
                "description": "Cancel one of your own orders (only while it is still `pending`).",
                "operationId": "cancel_order_api_v1_orders__order_id__cancel_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Order Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{order_id}/checkout": {
            "post": {
                "tags": [
                    "payments"
                ],
                "summary": "Checkout",
                "description": "Initiate payment for one of your `pending` orders with the chosen provider.\nReturns the provider-specific payload the client needs to finish paying\n(Stripe `client_secret`, or the bKash redirect URL).",
                "operationId": "checkout_api_v1_orders__order_id__checkout_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Order Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentInitiate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/StripeCheckoutResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/BkashCheckoutResponse"
                                        }
                                    ],
                                    "title": "Response Checkout Api V1 Orders  Order Id  Checkout Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/payments/bkash/callback": {
            "get": {
                "tags": [
                    "payments"
                ],
                "summary": "Bkash Callback",
                "operationId": "bkash_callback_api_v1_payments_bkash_callback_get",
                "parameters": [
                    {
                        "name": "paymentID",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Paymentid"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Status"
                        }
                    },
                    {
                        "name": "trxID",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Trxid"
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Signature"
                        }
                    },
                    {
                        "name": "apiVersion",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Apiversion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/payments/{payment_id}": {
            "get": {
                "tags": [
                    "payments"
                ],
                "summary": "Get Payment",
                "operationId": "get_payment_api_v1_payments__payment_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "payment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Payment Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BkashCheckoutResponse": {
                "properties": {
                    "bkash_url": {
                        "type": "string",
                        "title": "Bkash Url"
                    },
                    "payment_id": {
                        "type": "string",
                        "title": "Payment Id"
                    }
                },
                "type": "object",
                "required": [
                    "bkash_url",
                    "payment_id"
                ],
                "title": "BkashCheckoutResponse",
                "description": "Returned to the client so it can redirect the user into bKash's flow."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "OrderCreate": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/OrderItemCreate"
                        },
                        "type": "array",
                        "minItems": 1,
                        "title": "Items"
                    }
                },
                "type": "object",
                "required": [
                    "items"
                ],
                "title": "OrderCreate"
            },
            "OrderItemCreate": {
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "title": "Product Id"
                    },
                    "quantity": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "title": "Quantity"
                    }
                },
                "type": "object",
                "required": [
                    "product_id",
                    "quantity"
                ],
                "title": "OrderItemCreate",
                "description": "Client only sends product_id + quantity. price/subtotal are never\ntrusted from the client — OrderService looks up the live Product.price\nserver-side and computes both deterministically."
            },
            "OrderItemRead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "product_id": {
                        "type": "integer",
                        "title": "Product Id"
                    },
                    "quantity": {
                        "type": "integer",
                        "title": "Quantity"
                    },
                    "price": {
                        "type": "string",
                        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                        "title": "Price"
                    },
                    "subtotal": {
                        "type": "string",
                        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                        "title": "Subtotal"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "product_id",
                    "quantity",
                    "price",
                    "subtotal"
                ],
                "title": "OrderItemRead"
            },
            "OrderRead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    },
                    "total_amount": {
                        "type": "string",
                        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                        "title": "Total Amount"
                    },
                    "status": {
                        "$ref": "#/components/schemas/OrderStatus"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/OrderItemRead"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "user_id",
                    "total_amount",
                    "status",
                    "items",
                    "created_at"
                ],
                "title": "OrderRead"
            },
            "OrderStatus": {
                "type": "string",
                "enum": [
                    "pending",
                    "paid",
                    "canceled"
                ],
                "title": "OrderStatus"
            },
            "PaymentInitiate": {
                "properties": {
                    "order_id": {
                        "type": "integer",
                        "title": "Order Id"
                    },
                    "provider": {
                        "$ref": "#/components/schemas/PaymentProvider"
                    }
                },
                "type": "object",
                "required": [
                    "order_id",
                    "provider"
                ],
                "title": "PaymentInitiate",
                "description": "Body for POST /orders/{order_id}/checkout — user picks a provider."
            },
            "PaymentProvider": {
                "type": "string",
                "enum": [
                    "stripe",
                    "bkash"
                ],
                "title": "PaymentProvider"
            },
            "PaymentRead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "order_id": {
                        "type": "integer",
                        "title": "Order Id"
                    },
                    "provider": {
                        "$ref": "#/components/schemas/PaymentProvider"
                    },
                    "transaction_id": {
                        "type": "string",
                        "title": "Transaction Id"
                    },
                    "status": {
                        "$ref": "#/components/schemas/PaymentStatus"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "order_id",
                    "provider",
                    "transaction_id",
                    "status",
                    "created_at"
                ],
                "title": "PaymentRead"
            },
            "PaymentStatus": {
                "type": "string",
                "enum": [
                    "pending",
                    "success",
                    "failed"
                ],
                "title": "PaymentStatus"
            },
...
            "StripeCheckoutResponse": {
                "properties": {
                    "client_secret": {
                        "type": "string",
                        "title": "Client Secret"
                    },
                    "payment_intent_id": {
                        "type": "string",
                        "title": "Payment Intent Id"
                    }
                },
                "type": "object",
                "required": [
                    "client_secret",
                    "payment_intent_id"
                ],
                "title": "StripeCheckoutResponse",
                "description": "Returned to the client so it can complete the Stripe client-side confirm step."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    }
}